RW-8420 making return type integer

+++ api/src/main/java/org/pmiops/workbench/api/ConceptSetsController.java
@@ -45,7 +45,7 @@ public class ConceptSetsController implements ConceptSetsApiDelegate {
   }
 
   @Override
-  public ResponseEntity<Long> countConceptsInConceptSet(
+  public ResponseEntity<Integer> countConceptsInConceptSet(
       String workspaceNamespace, String workspaceId, Long conceptSetId) {
     workspaceAuthService.getWorkspaceEnforceAccessLevelAndSetCdrVersion(
         workspaceNamespace, workspaceId, WorkspaceAccessLevel.READER);
